Replace beacon.png with tiny gif
Instead of reading a PNG into memory and returning it, we can generate a valid GIF image on the fly, which reduces the response from 1.2kb to 42 bytes.

I've also added some headers so the email clients won't cache the beacon image, and everytime the user opens the email, it will be tracked.

I know the gif code looks ugly, but I'm using this technique on production for tracking emails and it definitely works.